backend/requirements.txt
# BrawlGPT Backend Dependencies
# Pin all versions for reproducible builds

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# HTTP client
requests==2.32.3
httpx==0.27.2

# Environment management
python-dotenv==1.0.1
pydantic-settings==2.0.3

# OpenAI SDK (for OpenRouter) - pinned to compatible version
openai==1.40.0

# Data validation
pydantic==2.9.2

# Database
sqlalchemy==2.0.27
asyncpg==0.29.0

# Rate limiting
slowapi==0.1.9

# Caching (for Phase 4)
cachetools==5.5.0
redis==5.0.1

# Authentication
passlib[argon2]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
